Tasks can be limited to a run context
By default all `Task` definitions will run in all context. This may not
be desirable for tasks that you want run only locally, or only on a
server (for example publishing artifacts).

A `task` now have the `limit_to` attribute. If set the task will be
limited to the run context, that can be passed to the runner via the
`--run-context` flag.

Possible values are `server` and `local`.